A lot of refactoring
