+++ Makefile
@@ -10,12 +10,12 @@ FTFYFLAG ="-DFORTIFY "
 FTFYLINTEXCL = --exclude=fortify.*
 endif
 
-SOURCES = src/bar.c src/options.c src/tiling_support.c src/windows.c src/dintwm.c src/commodity.c $(FTFYSRC)
+SOURCES = src/supplemental.c src/bar.c src/options.c src/tiling.c src/tiling_support.c src/windows.c src/dintwm.c src/commodity.c $(FTFYSRC)
 OBJECTS = $(SOURCES:.c=.o)
 CFLAGSSTRICT =-pedantic -std=c11 -ggdb3 $(FTFYFLAG)-O2 -Wall -Wextra -Wformat=2 -Wmissing-include-dirs -Winit-self -Wswitch-default -Wswitch-enum -Wunused-parameter -Wfloat-equal -Wundef -Wshadow -Wlarger-than-1000 -Wunsafe-loop-optimizations -Wbad-function-cast -Wcast-align -Wconversion -Wlogical-op -Waggregate-return -Wstrict-prototypes -Wold-style-definition -Wmissing-prototypes -Wmissing-declarations -Wpacked -Wpadded -Wredundant-decls -Wnested-externs -Wunreachable-code -Winline -Winvalid-pch -Wvolatile-register-var -Wstrict-aliasing=2 -Wstrict-overflow=2 -Wtraditional-conversion -Wwrite-strings -noixemul -save-temps=obj
 CC = m68k-amigaos-gcc
 CFLAGS =-std=c11 -Ofast -Wall -Wextra -s -fomit-frame-pointer -noixemul
-MAINHEADER = include/tiling.h include/windows.h include/dintwm.h include/dintwm_shared.h
+MAINHEADER = include/tiling_support.h include/windows.h include/dintwm.h include/dintwm_shared.h
 CONFHEADER = include/shared_defs.h include/commodity_types.h include/rawkey_combos.h include/commodity.h include/key_txts.h include/key_defs.h include/opt_defs.h
 EXTHEADERS = include/wbar_config.h $(FTFYHDRS)
 HEADERS = $(MAINHEADER) $(CONFHEADER) $(EXTHEADERS)
