+++ include/commodity_types.h
@@ -1,0 +1,121 @@
+// Copyright 2024 Rasmus Edgar
+#define TYPE_TILE "POPKEY_TILE"
+#define TYPE_HGRID "POPKEY_HGRID"
+#define TYPE_SPIRAL "POPKEY_SPIRAL"
+#define TYPE_DWINDLE "POPKEY_DWINDLE"
+#define TYPE_SWITCHF "POPKEY_SWITCHF"
+#define TYPE_SWITCHB "POPKEY_SWITCHB"
+#define TYPE_INCTOPGAP "POPKEY_INCTOPGAP"
+#define TYPE_INCBOTTOMGAP "POPKEY_INCBOTTOMGAP"
+#define TYPE_INCLEFTGAP "POPKEY_INCLEFTGAP"
+#define TYPE_INCRIGHTGAP "POPKEY_INCRIGHTGAP"
+#define TYPE_DECTOPGAP "POPKEY_DECTOPGAP"
+#define TYPE_DECBOTTOMGAP "POPKEY_DECBOTTOMGAP"
+#define TYPE_DECLEFTGAP "POPKEY_DECLEFTGAP"
+#define TYPE_DECRIGHTGAP "POPKEY_DECRIGHTGAP"
+#define TYPE_INCALLGAPS "POPKEY_INCALLGAPS"
+#define TYPE_DECALLGAPS "POPKEY_DECALLGAPS"
+#define TYPE_TILE_OFF "POPKEY_TILE_OFF"
+#define TYPE_KEY_CMD_0 "POPKEY_CMD_0"
+#define TYPE_KEY_CMD_1 "POPKEY_CMD_1"
+#define TYPE_KEY_CMD_2 "POPKEY_CMD_2"
+#define TYPE_KEY_CMD_3 "POPKEY_CMD_3"
+#define TYPE_KEY_CMD_4 "POPKEY_CMD_4"
+#define TYPE_KEY_CMD_5 "POPKEY_CMD_5"
+#define TYPE_KEY_CMD_6 "POPKEY_CMD_6"
+#define TYPE_KEY_CMD_7 "POPKEY_CMD_7"
+#define TYPE_KEY_CMD_8 "POPKEY_CMD_8"
+#define TYPE_KEY_CMD_9 "POPKEY_CMD_9"
+#define TYPE_KEY_WS_0 "POPKEY_WS_0"
+#define TYPE_KEY_WS_1 "POPKEY_WS_1"
+#define TYPE_KEY_WS_2 "POPKEY_WS_2"
+#define TYPE_KEY_WS_3 "POPKEY_WS_3"
+#define TYPE_KEY_WS_4 "POPKEY_WS_4"
+#define TYPE_KEY_WS_5 "POPKEY_WS_5"
+#define TYPE_KEY_CWS_0 "POPKEY_CWS_0"
+#define TYPE_KEY_CWS_1 "POPKEY_CWS_1"
+#define TYPE_KEY_CWS_2 "POPKEY_CWS_2"
+#define TYPE_KEY_CWS_3 "POPKEY_CWS_3"
+#define TYPE_KEY_CWS_4 "POPKEY_CWS_4"
+#define TYPE_KEY_CWS_5 "POPKEY_CWS_5"
+#define TYPE_KEY_CXM_EXIT "POPKEY_CXM_EXIT"
+#define TYPE_KEY_TAB_NEXT "POPKEY_TAB_NEXT"
+#define TYPE_TOPGAP "TOPGAP"
+#define TYPE_DEFAULT_TOPGAP "DEFAULT_TOPGAP"
+#define TYPE_BOTTOMGAP "BOTTOMGAP"
+#define TYPE_LEFTGAP "LEFTGAP"
+#define TYPE_RIGHTGAP "RIGHTGAP"
+#define TYPE_AUTO "AUTO"
+#define TYPE_AUTO_INTERVAL_DELAY "AUTO_INTERVAL_DELAY"
+#define TYPE_TILE_FACT "TILE_FACT"
+#define TYPE_GAP_CHANGE_VALUE "GAP_CHANGE_VALUE"
+#define TYPE_BAR "BAR"
+#define TYPE_BAR_HIDE_EMPTY "BAR_HIDE_EMPTY"
+#define TYPE_BAR_BG_COL "BAR_BG_COL"
+#define TYPE_BAR_FPW_COL "BAR_FPW_COL"
+#define TYPE_BAR_BPW_COL "BAR_BPW_COL"
+#define TYPE_BAR_FPCURW_COL "BAR_FPCURW_COL"
+#define TYPE_BAR_BPCURW_COL "BAR_BPCURW_COL"
+#define TYPE_BAR_FPTM_COL "BAR_FPTM_COL"
+#define TYPE_BAR_BPTM_COL "BAR_BPTM_COL"
+#define TYPE_BAR_FPTI_COL "BAR_FPTI_COL"
+#define TYPE_BAR_BPTI_COL "BAR_BPTI_COL"
+#define TYPE_BAR_FPSEP_ONE_COL "BAR_FPSEP_ONE_COL"
+#define TYPE_BAR_BPSEP_ONE_COL "BAR_BPSEP_ONE_COL"
+#define TYPE_BAR_FPSEP_TWO_COL "BAR_FPSEP_TWO_COL"
+#define TYPE_BAR_BPSEP_TWO_COL "BAR_BPSEP_TWO_COL"
+#define TYPE_BAR_TEXT_WS0 "BAR_TEXT_WS0"
+#define TYPE_BAR_TEXT_WS1 "BAR_TEXT_WS1"
+#define TYPE_BAR_TEXT_WS2 "BAR_TEXT_WS2"
+#define TYPE_BAR_TEXT_WS3 "BAR_TEXT_WS3"
+#define TYPE_BAR_TEXT_WS4 "BAR_TEXT_WS4"
+#define TYPE_BAR_TEXT_WS5 "BAR_TEXT_WS5"
+#define TYPE_BAR_TEXT_TILE "BAR_TEXT_TILE"
+#define TYPE_BAR_TEXT_GRID "BAR_TEXT_GRID"
+#define TYPE_BAR_TEXT_DWINDLE "BAR_TEXT_DWINDLE"
+#define TYPE_BAR_TEXT_SPIRAL "BAR_TEXT_SPIRAL"
+#define TYPE_BAR_TEXT_SEP_1 "BAR_TEXT_SEP_1"
+#define TYPE_BAR_TEXT_SEP_2 "BAR_TEXT_SEP_2"
+#define TYPE_BAR_TEXT_SPACE "BAR_TEXT_SPACE"
+#define TYPE_EXCL_WTYPE_0 "EXCL_WTYPE_0"
+#define TYPE_EXCL_WTYPE_1 "EXCL_WTYPE_1"
+#define TYPE_EXCL_WTYPE_2 "EXCL_WTYPE_2"
+#define TYPE_EXCL_WTYPE_3 "EXCL_WTYPE_3"
+#define TYPE_EXCL_WTYPE_4 "EXCL_WTYPE_4"
+#define TYPE_EXCL_WTYPE_5 "EXCL_WTYPE_5"
+#define TYPE_EXCL_WTYPE_6 "EXCL_WTYPE_6"
+#define TYPE_EXCL_WTYPE_7 "EXCL_WTYPE_7"
+#define TYPE_EXCL_WTYPE_8 "EXCL_WTYPE_8"
+#define TYPE_EXCL_WTYPE_9 "EXCL_WTYPE_9"
+#define TYPE_INCL_WTYPE_0 "INCL_WTYPE_0"
+#define TYPE_INCL_WTYPE_1 "INCL_WTYPE_1"
+#define TYPE_INCL_WTYPE_2 "INCL_WTYPE_2"
+#define TYPE_INCL_WTYPE_3 "INCL_WTYPE_3"
+#define TYPE_INCL_WTYPE_4 "INCL_WTYPE_4"
+#define TYPE_INCL_WTYPE_5 "INCL_WTYPE_5"
+#define TYPE_INCL_WTYPE_6 "INCL_WTYPE_6"
+#define TYPE_INCL_WTYPE_7 "INCL_WTYPE_7"
+#define TYPE_INCL_WTYPE_8 "INCL_WTYPE_8"
+#define TYPE_INCL_WTYPE_9 "INCL_WTYPE_9"
+#define TYPE_CONLINE_0 "CONLINE_0"
+#define TYPE_CONLINE_1 "CONLINE_1"
+#define TYPE_CONLINE_2 "CONLINE_2"
+#define TYPE_CONLINE_3 "CONLINE_3"
+#define TYPE_CONLINE_4 "CONLINE_4"
+#define TYPE_CONLINE_5 "CONLINE_5"
+#define TYPE_CONLINE_6 "CONLINE_6"
+#define TYPE_CONLINE_7 "CONLINE_7"
+#define TYPE_CONLINE_8 "CONLINE_8"
+#define TYPE_CONLINE_9 "CONLINE_9"
+#define TYPE_CMD_0 "CMD_0"
+#define TYPE_CMD_1 "CMD_1"
+#define TYPE_CMD_2 "CMD_2"
+#define TYPE_CMD_3 "CMD_3"
+#define TYPE_CMD_4 "CMD_4"
+#define TYPE_CMD_5 "CMD_5"
+#define TYPE_CMD_6 "CMD_6"
+#define TYPE_CMD_7 "CMD_7"
+#define TYPE_CMD_8 "CMD_8"
+#define TYPE_CMD_9 "CMD_9"
+#define TYPE_INFO_OFF "INFO_OFF"
+#define TYPE_VWS_ON "VWS_ON"
