+++ include/supplemental.h
@@ -1,0 +1,2 @@
+// Copyright 2021 Rasmus Edgar
+#include <dos/dostags.h>
